new cache WIP

+++ src/libertem/io/cache/cache.py
@@ -1,7 +1,7 @@
 import glob
 import os
 
-from libertem.io.cache.stats import CacheStats, CacheItem, OrphanItem
+from libertem.io.cache.stats import CacheStats, CacheItem
 from libertem.io.cache.strategy import CacheStrategy
 
 
@@ -27,20 +27,35 @@ class Cache:
         Make place for `size` bytes which will be used
         by the dataset identified by the `cache_key`.
         """
+
+        if self.strategy.sufficient_space_for(size, self._stats):
+            return []
+
+        evicted = []
         with self._stats:
-            victims = self.strategy.get_victim_list(cache_key, size, self._stats)
-            for cache_item in victims:
-                # if it has been deleted by the user, we don't care and just remove
-                # the record from the database:
+            candidates = self.strategy.get_victim_list(cache_key, size, self._stats)
+            for cache_item in candidates:
+                # if item has been deleted by the user or another task
+                # don't add to evicted list, as the space may already
+                # be assumed free by another task! do record eviction,
+                # to not leave files behind in any case
+                self._stats.record_eviction(cache_item)
                 if os.path.exists(cache_item.path):
                     os.unlink(cache_item.path)
-                self._stats.record_eviction(cache_item)
+                else:
+                    continue
+                evicted.append(cache_item)
+
+                if sum(i.size for i in evicted) >= size:
+                    break
+        return evicted
 
     def collect_orphans(self, base_path: str):
         """
         Check the filesystem structure and record all partitions
         that are missing in the db as orphans, to be deleted on demand.
         """
+        raise NotImplementedError("to be reimplemented")
         # the structure here is: {base_path}/{dataset_cache_key}/parts/*
         orphans = []
         with self._stats:
